requirements.txt
Flask==3.0.3
gunicorn
python-dotenv
apscheduler
psycopg2-binary
SQLAlchemy
